python-manager/modules/spell-grammar-checker/requirements.txt
pyspellchecker==0.8.1
language-tool-python==2.8
lxml==5.1.0
